x64警告対応 CDialog::DoModalの戻り値をintにキャストする

+++ sakura_core/dlg/CDlgJump.cpp
@@ -64,7 +64,7 @@ int CDlgJump::DoModal(
 	LPARAM		lParam
 )
 {
-	return CDialog::DoModal( hInstance, hwndParent, IDD_JUMP, lParam );
+	return (int)CDialog::DoModal( hInstance, hwndParent, IDD_JUMP, lParam );
 }
 
 // From Here Oct. 6, 2000 JEPRO added 行番号入力ボックスにスピンコントロールを付けるため
